Fixes #18837 - Delete reports when host is missing

+++ app/models/foreman_openscap/arf_report.rb
@@ -164,11 +164,15 @@ module ForemanOpenscap
     end
 
     def destroy
-      begin
-        openscap_proxy_api.destroy_report(self, ForemanOpenscap::Helper::find_name_or_uuid_by_host(host))
-      rescue Foreman::Exception => e
-        logger.error "Failed to delete report with id #{id} from proxy, cause: #{e.message}"
-        logger.debug e.backtrace.join("\n\t")
+      if host
+        begin
+          openscap_proxy_api.destroy_report(self, ForemanOpenscap::Helper::find_name_or_uuid_by_host(host))
+        rescue Foreman::Exception => e
+          logger.error "Failed to delete report with id #{id} from proxy, cause: #{e.message}"
+          logger.debug e.backtrace.join("\n\t")
+        end
+      else
+        logger.error "Failed to delete report with id #{id} from proxy, no host associated with report"
       end
       super
     end
